tec: Move Importations controller
It wasn't coherent with other controllers such as Collections which is
not under the `collections/` directory.

I prefer to have all the main controllers directly under `controllers/`
because it avoids to have some of these under a subdirectory and not
some others.
